refactor: Receipt, ReceiptDetail, user 패키지에 리팩토링
