+++ src/main/java/uos/uos25/user/receipt/repository/ReceiptDetailRepository.java
@@ -1,11 +1,11 @@
-package uos.uos25.receipt.repository;
+package uos.uos25.user.receipt.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.receipt.entity.ReceiptDetail;
+import uos.uos25.user.receipt.entity.ReceiptDetail;
 
 @Repository
 public interface ReceiptDetailRepository extends JpaRepository<ReceiptDetail, Long> {
