+++ src/main/java/uos/uos25/user/receipt/repository/ReceiptRepository.java
@@ -1,11 +1,11 @@
-package uos.uos25.receipt.repository;
+package uos.uos25.user.receipt.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.receipt.entity.Receipt;
+import uos.uos25.user.receipt.entity.Receipt;
 
 @Repository
 public interface ReceiptRepository extends JpaRepository<Receipt, Long> {
